+++ src/main/scala/edu/ie3/osmogrid/guardian/run/StopSupport.scala
@@ -8,6 +8,7 @@ package edu.ie3.osmogrid.guardian.run
 
 import org.apache.pekko.actor.typed.scaladsl.ActorContext
 import edu.ie3.osmogrid.io.input.InputTerminate
+import edu.ie3.osmogrid.io.output.StopListener
 import edu.ie3.osmogrid.lv.LvTerminate
 import edu.ie3.osmogrid.mv.MvTerminate
 
@@ -28,16 +29,18 @@ trait StopSupport {
   protected def stopChildren(
       runId: UUID,
       childReferences: ChildReferences,
-      ctx: ActorContext[RunRequest],
+      ctx: ActorContext[Messages],
   ): StoppingData = {
     childReferences.lvCoordinator.foreach(_ ! LvTerminate)
     childReferences.mvCoordinator.foreach(_ ! MvTerminate)
     childReferences.inputDataProvider ! InputTerminate
+    childReferences.resultListeners.foreach(_ ! StopListener)
 
     StoppingData(
       runId,
       inputDataProviderTerminated = false,
       resultListenerTerminated = false,
+      childReferences.poiParser.map(_ => false),
       childReferences.lvCoordinator.map(_ => false),
       childReferences.mvCoordinator.map(_ => false),
     )
@@ -61,6 +64,10 @@ trait StopSupport {
       stoppingData.copy(inputDataProviderTerminated = true)
     case ResultEventListenerDied =>
       stoppingData.copy(resultListenerTerminated = true)
+    case PoiParserDied =>
+      stoppingData.copy(poiParserTerminated =
+        stoppingData.poiParserTerminated.map(_ => true)
+      )
     case LvCoordinatorDied =>
       stoppingData.copy(lvCoordinatorTerminated =
         stoppingData.lvCoordinatorTerminated.map(_ => true)
@@ -89,7 +96,7 @@ trait StopSupport {
       runId: UUID,
       childReferences: ChildReferences,
       watchMsg: RunWatch,
-      ctx: ActorContext[RunRequest],
+      ctx: ActorContext[Messages],
   ): StoppingData = {
     (stopChildren(runId, childReferences, ctx), watchMsg) match {
       case (stoppingData, InputDataProviderDied) =>
@@ -102,6 +109,13 @@ trait StopSupport {
           s"One of the result listener for run $runId unexpectedly died. Start coordinated shut down phase for this run."
         )
         stoppingData.copy(resultListenerTerminated = true)
+      case (stoppingData, PoiParserDied) =>
+        ctx.log.warn(
+          s"POI Parser for run $runId unexpectedly died. Start coordinated shut down phase for this run."
+        )
+        stoppingData.copy(poiParserTerminated =
+          stoppingData.poiParserTerminated.map(_ => true)
+        )
       case (stoppingData, LvCoordinatorDied) =>
         ctx.log.warn(
           s"Lv coordinator for run $runId unexpectedly died. Start coordinated shut down phase for this run."