+++ src/main/scala/edu/ie3/osmogrid/io/output/ResultCsvSink.scala
@@ -6,6 +6,7 @@
 
 package edu.ie3.osmogrid.io.output
 
+import edu.ie3.datamodel.io.csv.BufferedCsvWriter
 import edu.ie3.datamodel.io.naming.{
   DefaultDirectoryHierarchy,
   EntityPersistenceNamingStrategy,
@@ -13,8 +14,9 @@ import edu.ie3.datamodel.io.naming.{
   FlatDirectoryHierarchy,
 }
 import edu.ie3.datamodel.io.sink.CsvFileSink
+import edu.ie3.osmogrid.poi.PoiElement
 
-import java.nio.file.Path
+import java.nio.file.{Files, Path}
 import java.util.UUID
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
@@ -36,6 +38,32 @@ final case class ResultCsvSink(
         new FlatDirectoryHierarchy(),
     ),
     csvSeparator,
+  )
+
+  private val poiSink = {
+    val path = saveFolderPath.resolve("poi")
+    Files.createDirectories(path)
+    val writer = new BufferedCsvWriter(
+      path.resolve("poi.csv"),
+      Array("id", "size", "lat", "lon"),
+      csvSeparator,
+      false,
+    )
+    writer.writeFileHeader()
+    writer
+  }
+
+  private given Conversion[PoiElement, java.util.Map[String, String]] = poi => {
+    val map = new java.util.HashMap[String, String](4)
+    map.put("id", s"${poi.poiType}" + s"_${poi.id}")
+    map.put("size", poi.size.toString)
+    map.put("lat", poi.lat.toString)
+    map.put("lon", poi.lon.toString)
+    map
+  }
+
+  override def handlePOIs(pois: Iterable[PoiElement]): Future[Unit] = Future(
+    pois.foreach(poiSink.write(_))
   )
 
   def handleResult(