+++ src/main/scala/edu/ie3/osmogrid/io/output/ResultSink.scala
@@ -6,6 +6,8 @@
 
 package edu.ie3.osmogrid.io.output
 
+import edu.ie3.osmogrid.poi.PoiElement
+
 import scala.concurrent.Future
 
 /** Trait that should be mixed into each sink in [[edu.ie3.osmogrid.io.output]]
@@ -13,6 +15,8 @@ import scala.concurrent.Future
   * sink
   */
 trait ResultSink {
+
+  def handlePOIs(pois: Iterable[PoiElement]): Future[Unit]
 
   def handleResult(gridResult: GridResult): Future[Unit]