+++ src/main/scala/edu/ie3/osmogrid/lv/LvCoordinator.scala
@@ -11,6 +11,7 @@ import org.apache.pekko.actor.typed.{ActorRef, Behavior, PostStop}
 import edu.ie3.osmogrid.ActorStopSupportStateless
 import edu.ie3.osmogrid.cfg.OsmoGridConfig
 import edu.ie3.osmogrid.graph.OsmGraph
+import edu.ie3.osmogrid.io.input.FilterType.LV
 import edu.ie3.osmogrid.io.input.{
   BoundaryAdminLevel,
   InputDataEvent,
@@ -90,19 +91,9 @@ object LvCoordinator extends ActorStopSupportStateless {
           ctx.log.debug("Request input data")
 
           /* Ask for OSM data */
-          val filter = stateData.cfg.osm.filter
-            .map(cfg =>
-              LvFilter(
-                cfg.building.toSet,
-                cfg.highway.toSet,
-                cfg.landuse.toSet,
-              )
-            )
-            .getOrElse(LvFilter())
-
           stateData.inputDataProvider ! ReqOsm(
             replyTo = stateData.msgAdapters.inputDataProvider,
-            filter = filter,
+            filterType = LV,
           )
           /* Ask for grid asset data */
           stateData.inputDataProvider ! ReqAssetTypes(